requirements.txt
astor==0.8.1
bids-validator==1.11.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
fonttools==4.39.4
formulaic==0.5.2
greenlet==2.0.2
idna==3.4
interface-meta==1.3.0
joblib==1.2.0
kaleido==0.2.1
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
nibabel==5.1.0
nilearn==0.10.1
nireports==23.1.0
jinja2==3.1.2
num2words==0.5.12
numpy==1.24.3
packaging==23.1
pandas==2.0.1
patsy==0.5.3
PeakUtils==1.3.4
Pillow==9.5.0
plotly==5.14.1
pybids==0.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.14
statsmodels==0.14.0
tenacity==8.2.2
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
xmlformatter==0.2.4
urllib3==2.0.2
uuid==1.30
wrapt==1.15.0
pytest==7.4.3
